% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filter_by_type}
\alias{filter_by_type}
\title{Filtrer le dataset par type de bien}
\usage{
filter_by_type(data, type)
}
\arguments{
\item{data}{Le jeu de données}

\item{type}{Le type de bien}
}
\value{
Un data frame filtré
}
\description{
Filtrer le dataset par type de bien
}
