% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{mean_price_by_group}
\alias{mean_price_by_group}
\title{Calcul du prix moyen par arrondissement}
\usage{
mean_price_by_group(data, groupe)
}
\arguments{
\item{data}{Le jeu de données}

\item{groupe}{La variable de groupement}
}
\value{
Un data frame avec le prix moyen par groupe
}
\description{
Calcul du prix moyen par arrondissement
}
