% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot_mean_price}
\alias{plot_mean_price}
\title{Graphique du prix moyen par arrondissement}
\usage{
plot_mean_price(data, groupe)
}
\arguments{
\item{data}{Le jeu de données}

\item{groupe}{La variable de groupement}
}
\value{
Un graphique ggplot2
}
\description{
Graphique du prix moyen par arrondissement
}
